Split widget code into separate files
