+++ pynicotine/gtkgui/downloads.py
@@ -28,9 +28,9 @@ import re
 
 from gi.repository import Gtk
 
-from pynicotine.gtkgui.dialogs import option_dialog
 from pynicotine.gtkgui.transferlist import TransferList
 from pynicotine.gtkgui.utils import open_file_path
+from pynicotine.gtkgui.widgets.messagedialogs import option_dialog
 from pynicotine.logfacility import log
