+++ pynicotine/gtkgui/statistics.py
@@ -20,9 +20,9 @@ import os
 
 from gi.repository import Gtk
 
-from pynicotine.gtkgui.dialogs import option_dialog
-from pynicotine.gtkgui.utils import human_size
 from pynicotine.gtkgui.utils import load_ui_elements
+from pynicotine.gtkgui.widgets.messagedialogs import option_dialog
+from pynicotine.utils import human_size
 
 
 class Statistics:
