+++ pynicotine/gtkgui/transferlist.py
@@ -32,21 +32,21 @@ from gi.repository import GObject
 from gi.repository import Gtk
 
 from pynicotine.gtkgui.fileproperties import FileProperties
-from pynicotine.gtkgui.utils import collapse_treeview
 from pynicotine.gtkgui.utils import copy_file_url
-from pynicotine.gtkgui.utils import human_size
-from pynicotine.gtkgui.utils import human_speed
-from pynicotine.gtkgui.utils import initialise_columns
 from pynicotine.gtkgui.utils import keyval_to_hardware_keycode
 from pynicotine.gtkgui.utils import load_ui_elements
-from pynicotine.gtkgui.utils import PopupMenu
-from pynicotine.gtkgui.utils import save_columns
-from pynicotine.gtkgui.utils import select_user_row_iter
-from pynicotine.gtkgui.utils import set_treeview_selected_row
-from pynicotine.gtkgui.utils import show_file_path_tooltip
 from pynicotine.gtkgui.utils import triggers_context_menu
-from pynicotine.gtkgui.utils import update_widget_visuals
+from pynicotine.gtkgui.widgets.popupmenu import PopupMenu
+from pynicotine.gtkgui.widgets.theme import update_widget_visuals
+from pynicotine.gtkgui.widgets.treeview import collapse_treeview
+from pynicotine.gtkgui.widgets.treeview import initialise_columns
+from pynicotine.gtkgui.widgets.treeview import save_columns
+from pynicotine.gtkgui.widgets.treeview import select_user_row_iter
+from pynicotine.gtkgui.widgets.treeview import set_treeview_selected_row
+from pynicotine.gtkgui.widgets.treeview import show_file_path_tooltip
 from pynicotine.transfers import Transfer
+from pynicotine.utils import human_size
+from pynicotine.utils import human_speed
 
 
 class TransferList:
