+++ pynicotine/gtkgui/uploads.py
@@ -24,9 +24,9 @@
 
 import os
 
-from pynicotine.gtkgui.dialogs import option_dialog
 from pynicotine.gtkgui.transferlist import TransferList
 from pynicotine.gtkgui.utils import open_file_path
+from pynicotine.gtkgui.widgets.messagedialogs import option_dialog
 
 
 class Uploads(TransferList):
