+++ test/unit/search/test_search.py
@@ -22,18 +22,27 @@ import pytest
 from collections import deque
 
 from pynicotine.config import Config
+from pynicotine.i18n import apply_translation
 from pynicotine.search import Search
 
 SEARCH_TEXT = '70 gwen "test" -mp3 -nothanks a:b;c+d +++---}[ [[ @@ auto -no yes'
 SEARCH_MODE = 'global'
 
 
+@pytest.fixture(scope="module", autouse=True)
+def setup():
+    # Setting gettext and locale
+    apply_translation()
+
+
 @pytest.fixture
 def config():
-    data_path = os.path.dirname(os.path.realpath(__file__))
-    config_path = os.path.join(data_path, "temp_config")
+    config = Config()
+    config.data_dir = os.path.dirname(os.path.realpath(__file__))
+    config.filename = os.path.join(config.data_dir, "temp_config")
 
-    return Config(config_path, data_path)
+    config.load_config()
+    return config
 
 
 @pytest.fixture
